更改接口设计，完善mysql用户权限分配（需要XA_RECOVER_ADMIN），初步测试新建/更新/回滚笔记接口

+++ Note_back/src/main/java/com/riptidez/notebackend/note/entity/Note.java
@@ -7,7 +7,7 @@ package com.riptidez.notebackend.note.entity;
 public class Note {
     private Long id;               // Note.id
     private Long userId;           // Note.user_id（指向 User）
-    private Long currentHistoryId; // Note.current_history_id（指向 NoteHistory）
+    private Integer currentHistoryId; // Note.current_history_id（指向 NoteHistory）
 
     public Long getId() {
         return id;
@@ -25,11 +25,11 @@ public class Note {
         this.userId = userId;
     }
 
-    public Long getCurrentHistoryId() {
+    public Integer getCurrentHistoryId() {
         return currentHistoryId;
     }
 
-    public void setCurrentHistoryId(Long currentHistoryId) {
+    public void setCurrentHistoryId(Integer currentHistoryId) {
         this.currentHistoryId = currentHistoryId;
     }
